generic validator base class
